+++ tests/unit/test_experiments/test_simulation_with_experiment.py
@@ -765,6 +765,28 @@ class TestSimulationExperiment(TestCase):
         # Check that there are only 3 built models (unique steps + padding rest)
         self.assertEqual(len(sim.op_conds_to_built_models), 3)
 
+    def test_experiment_custom_termination(self):
+        def neg_stoich_cutoff(variables):
+            return variables["Negative electrode stoichiometry"] - 0.5
+
+        neg_stoich_termination = pybamm.step.CustomTermination(
+            name="Negative stoichiometry cut-off", event_function=neg_stoich_cutoff
+        )
+
+        model = pybamm.lithium_ion.SPM()
+        experiment = pybamm.Experiment(
+            [pybamm.step.c_rate(1, termination=neg_stoich_termination)]
+        )
+        sim = pybamm.Simulation(model, experiment=experiment)
+        sol = sim.solve(calc_esoh=False)
+        self.assertEqual(
+            sol.cycles[0].steps[0].termination,
+            "event: Negative stoichiometry cut-off [experiment]",
+        )
+
+        neg_stoich = sol["Negative electrode stoichiometry"].data
+        self.assertAlmostEqual(neg_stoich[-1], 0.5, places=4)
+
 
 if __name__ == "__main__":
     print("Add -v for more debug output")